+++ storage/pkg/flagset/storagemetadata.go
@@ -106,6 +106,7 @@ func StorageMetadata(cfg *config.Config) []cli.Flag {
 	f = append(f, DriverOwnCloudWithConfig(cfg)...)
 	f = append(f, DriverOCISWithConfig(cfg)...)
 	f = append(f, DriverS3NGWithConfig(cfg)...)
+
 	return f
 
 }